Moved store getters to dedicated file Moved search / sort logic from shootList utils to store/shoot getter
